authorization

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Models\Post;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Str;
 use SweetAlert2\Laravel\Swal;
@@ -70,6 +71,7 @@ class PostController extends Controller
      */
     public function update(Request $request, Post $post)
     {
+        Gate::authorize('update',$post);
         $data = $request->validate(['image' => 'nullable|mimes:png,jpg,gif',
             'description' => 'required']);
         $request->hasFile('image') ? $data['image'] = $request->file('image')->store('posts', 'public') : '';
@@ -83,9 +85,10 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        Gate::authorize('delete',$post);
         $post->delete();
         Storage::disk('public')->delete($post->image);
-        return redirect(route('welcome'));
+        return redirect(route('user_profile', $post->owner->userName));
     }
 
     public function explore(){
